Remove CarouselContainer from common to SelectVideo.tsx
